+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [0.2.2] - 2021-12-22
+### Added
+- Ability to reshuffle to "Result Page"
 ### Fixes
 - Temporary fix for the possibility of assigning 1st santa to himself when the number of santas is odd number